* changed pk + fk in db to int(10)

+++ config/db/tables.sql
@@ -6,9 +6,9 @@ CREATE TABLE `ext_calendar_event` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL DEFAULT '0',
-  `id_person_create` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id_person_create` int(10) unsigned NOT NULL DEFAULT '0',
   `deleted` tinyint(1) unsigned NOT NULL DEFAULT '0',
-  `id_project` smallint(6) NOT NULL DEFAULT '0',
+  `id_project` int(10) NOT NULL DEFAULT '0',
   `id_task` int(10) unsigned NOT NULL,
   `eventtype` tinyint(2) unsigned NOT NULL DEFAULT '0',
   `title` varchar(255) NOT NULL,
@@ -28,10 +28,10 @@ CREATE TABLE `ext_calendar_event` (
 --
 
 CREATE TABLE `ext_calendar_holiday` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL DEFAULT '0',
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(1) NOT NULL DEFAULT '0',
   `date` int(11) NOT NULL DEFAULT '0',
   `title` varchar(48) NOT NULL,
@@ -47,10 +47,10 @@ CREATE TABLE `ext_calendar_holiday` (
 --
 
 CREATE TABLE `ext_calendar_holidayset` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL DEFAULT '0',
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `deleted` tinyint(2) NOT NULL,
   `title` varchar(32) NOT NULL,
   `description` varchar(128) NOT NULL,
@@ -64,7 +64,7 @@ CREATE TABLE `ext_calendar_holidayset` (
 --
 
 CREATE TABLE `ext_calendar_mm_holiday_holidayset` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `id_holiday` int(10) unsigned NOT NULL DEFAULT '0',
   `id_holidayset` int(10) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
@@ -79,9 +79,9 @@ CREATE TABLE `ext_calendar_mm_holiday_holidayset` (
 --
 
 CREATE TABLE `ext_calendar_mm_event_person` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `id_event` int(10) unsigned NOT NULL DEFAULT '0',
-  `id_person` smallint(5) unsigned NOT NULL,
+  `id_person` int(10) unsigned NOT NULL,
   `is_acknowledged` tinyint(2) NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `event` (`id_event`),
